2004-04-14  Andreas Nahr <ClassDevelopment@A-SoftTech.com>
	* win32enum: made Msg internal
	* Control.cs
	* NativeWindow.cs
	* ScrollableControl.cs
	* Form.cs
	* Application.cs
	* ButtonBase.cs
	* ListBox.cs
	* CheckedListBox.cs
	* ComboBox.cs
	* TextBoxBase.cs
	* DateTimePicker.cs
	* GroupBox.cs
	* ScrollBar.cs
	* Label.cs
	* ListView.cs
	* StatusBar.cs
	* TabControl.cs
	* TrackBar.cs
	* UpDownBase.cs
	* Message.cs: Changes that were neccessary because of
	  making Msg internal

svn path=/trunk/mcs/; revision=25503

+++ mcs/class/System.Windows.Forms/System.Windows.Forms/Application.cs
@@ -1,5 +1,5 @@
 //
-// System.Windows.Forms.Application
+// System.Windows.Forms.Application.cs
 //
 // Author:
 //   stubbed out by Jaak Simm (jaaksimm@firm.ee)
@@ -316,7 +316,7 @@ namespace System.Windows.Forms
 
 				Message message = new Message ();
 				message.HWnd = msg.hwnd;
-				message.Msg = msg.message;//
+				message.Msg = (int) msg.message;//
 				message.WParam = msg.wParam;
 				message.LParam = msg.lParam;
 
@@ -386,7 +386,7 @@ namespace System.Windows.Forms
 
 				Message message = new Message ();
 				message.HWnd = msg.hwnd;
-				message.Msg = msg.message;
+				message.Msg = (int) msg.message;
 				message.WParam = msg.wParam;
 				message.LParam = msg.lParam;
